refactor: move miscelaneous epoch APIs to epoch manager adapter
cc #6910
